Add custom-set file and folder and change load-path style
To prevent things like these being stored in the init.el file we are instead
storing this in the settings.el file.

Addition of snippets folder to the ignore file until we have the subtree in place to add in the snippets as sub folders.

Also update the emacs dir assumption.